Updated CafeLib.Blazor fixed DroneLander Sample.
